Move self.postMessage transfer tests to standalone file
worker.test.ts has pre-existing timeout-sensitive failures on slow ASAN
builds that are unrelated to this change. Keep the new transfer-list tests
in their own file in the same directory so they can be verified
independently.
